Switch from Jekyll to Eleventy.
I don't have anything against Jekyll or Ruby, but the github-pages gem
uses seriously outdated versions of both. As a result, local development
(especially testing actions) was simply painful enough to convince me to
drop it entirely. Freed of the default/builtin SSG for Pages, I switched
to Eleventy rather than an up-to-date Jekyll mostly because I've been
wanting to try it.